Wire up mobile comment timestamp linking
